requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
aiohttp==3.9.0
requests==2.31.0
beautifulsoup4==4.12.0
lxml==4.9.0
selenium==4.15.0
webdriver-manager==4.0.0
httpx==0.25.0
